+++ packages/cli/doc-core/src/theme-default/layout/DocLayout/docComponents/hr.tsx
@@ -1,0 +1,10 @@
+import { ComponentProps } from 'react';
+
+export const Hr = (props: ComponentProps<'hr'>) => {
+  return (
+    <hr
+      {...props}
+      className="my-4 border-t bordder-solid border-divider-light"
+    />
+  );
+};