+++ packages/cli/doc-core/src/theme-default/layout/DocLayout/docComponents/link.tsx
@@ -1,0 +1,6 @@
+import { ComponentProps } from 'react';
+import styles from './index.module.scss';
+
+export const A = (props: ComponentProps<'a'>) => {
+  return <a {...props} className={`${styles.link} ${props.className}`} />;
+};