+++ packages/builder/plugin-swc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/builder-plugin-swc
 
+## 2.9.0
+
+### Patch Changes
+
+- @modern-js/builder-shared@2.9.0
+- @modern-js/utils@2.9.0
+
 ## 2.8.0
 
 ### Minor Changes