+++ packages/cli/babel-preset-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/babel-preset-app
 
+## 2.9.0
+
+### Patch Changes
+
+- @modern-js/babel-preset-base@2.9.0
+- @modern-js/utils@2.9.0
+
 ## 2.8.0
 
 ### Patch Changes