+++ packages/cli/babel-preset-base/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/babel-preset-base
 
+## 2.9.0
+
+### Patch Changes
+
+- @modern-js/utils@2.9.0
+
 ## 2.8.0
 
 ### Patch Changes