+++ packages/document/module-doc/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/module-tools-docs
 
+## 2.9.0
+
+### Patch Changes
+
+- 540205e909: docs(module-tools-doc): format content
+  docs(module-tools-doc): 格式化内容
+  - @modern-js/doc-tools@2.9.0
+  - @modern-js/doc-plugin-auto-sidebar@2.9.0
+
 ## 2.8.0
 
 ### Patch Changes