+++ packages/generator/generators/entry-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/entry-generator
 
+## 3.0.11
+
 ## 3.0.10
 
 ## 3.0.9