+++ packages/server/plugin-egg/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/plugin-egg
 
+## 2.9.0
+
+### Patch Changes
+
+- @modern-js/bff-core@2.9.0
+- @modern-js/bff-runtime@2.9.0
+- @modern-js/utils@2.9.0
+
 ## 2.8.0
 
 ### Patch Changes