+++ packages/solutions/doc-tools/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @modern-js/doc-tools
 
+## 2.9.0
+
+### Patch Changes
+
+- Updated dependencies [c61aabfbfe]
+- Updated dependencies [226ba4b061]
+- Updated dependencies [abb0658221]
+- Updated dependencies [88328d914b]
+- Updated dependencies [dd56401486]
+- Updated dependencies [8f7d293462]
+  - @modern-js/doc-core@2.9.0
+  - @modern-js/core@2.9.0
+  - @modern-js/utils@2.9.0
+
 ## 2.8.0
 
 ### Patch Changes