+++ scripts/check-changeset/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @scripts/check-changeset
 
+## 2.9.0
+
 ## 2.8.0