+++ tests/e2e/builder/cases/dev/basic/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @e2e/builder-dev-basic
 
+## 2.9.0
+
 ## 2.8.0