+++ tests/e2e/builder/cases/inspect-config/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @e2e/webpack-builder-inspect-config
 
+## 2.9.0
+
 ## 2.8.0