+++ tests/e2e/builder/cases/typescript/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @e2e/webpack-builder-typescript
 
+## 2.9.0
+
 ## 2.8.0