+++ tests/integration/load-config/fixtures/async-config-function/CHANGELOG.md
@@ -1,3 +1,5 @@
 # integration-config-async-config-test
 
+## 2.9.0
+
 ## 2.8.0