+++ tests/integration/load-config/fixtures/config-function-params/CHANGELOG.md
@@ -1,3 +1,5 @@
 # integration-config-function-params
 
+## 2.9.0
+
 ## 2.8.0